+++ ui_display.py
@@ -1,0 +1,1224 @@
+"""
+Manifest Alerts V2 - SpaceX-Style Mission Control Display
+Modern, professional TV display for warehouse operations
+"""
+
+import sys
+import json
+import os
+from datetime import datetime, timedelta
+from PyQt6.QtWidgets import (QWidget, QVBoxLayout, QHBoxLayout, QLabel, 
+                             QGridLayout, QFrame, QPushButton,
+                             QMessageBox, QScrollArea, QApplication, QDialog,
+                             QLineEdit, QDialogButtonBox, QFormLayout)
+from PyQt6.QtGui import QFont, QIcon
+from PyQt6.QtCore import Qt, QTimer
+
+# Import with error handling
+try:
+    from scheduler import get_manifest_status
+    from data_manager import load_config
+    from settings_manager import get_settings_manager
+except ImportError:
+    # Fallback for testing
+    def get_manifest_status(time_str, now):
+        return "OPEN"
+    def load_config():
+        return {"manifests": [
+            {"time": "08:00", "carriers": ["DHL", "UPS"]},
+            {"time": "12:00", "carriers": ["FedEx", "USPS"]},
+            {"time": "16:00", "carriers": ["DPD", "TNT"]}
+        ]}
+    def get_settings_manager():
+        class DummySettings:
+            def get_acknowledgments_path(self):
+                return "ack.json"
+        return DummySettings()
+
+
+class StatusCard(QFrame):
+    """Modern status card widget"""
+    
+    def __init__(self, time_str, status="OPEN", parent_display=None):
+        super().__init__()
+        self.time_str = time_str
+        self.status = status
+        self.manifests = []
+        self.parent_display = parent_display  # Reference to main display for acknowledgments
+        
+        self.setMinimumSize(1200, 80)  # Reduced minimum height significantly
+        self.setFrameStyle(QFrame.Shape.Box)
+        self.setup_ui()
+        self.update_styling()
+        
+        # Remove card-level hover tracking - we'll handle individual carrier hover instead
+        
+        # Make card clickable (but we'll override for time header double-click)
+        self.mousePressEvent = self.card_clicked
+    
+    def setup_ui(self):
+        layout = QHBoxLayout()  # Main horizontal layout
+        layout.setContentsMargins(10, 3, 10, 3)  # Even smaller margins - reduced from 15,5 to 10,3
+        layout.setSpacing(15)  # Reduced spacing between sections
+        
+        # Left side - Time and Status (fixed width for consistent alignment)
+        time_status_layout = QVBoxLayout()
+        time_status_layout.setSpacing(0)  # No spacing in time section
+        time_status_layout.setContentsMargins(0, 0, 0, 0)  # No margins
+        
+        # Combined time and status header
+        self.time_status_label = QLabel(f"{self.time_str} - {self.status}")
+        self.time_status_label.setAlignment(Qt.AlignmentFlag.AlignLeft | Qt.AlignmentFlag.AlignTop)
+        self.time_status_label.setFont(QFont("Segoe UI", 24, QFont.Weight.Bold))  # Increased from 22 to 24
+        self.time_status_label.setFixedWidth(280)  # Consistent fixed width for alignment
+        self.time_status_label.mouseDoubleClickEvent = self.time_header_double_clicked
+        # Set up hover effects with mouse events
+        self.time_status_label.enterEvent = self.time_header_hover_enter
+        self.time_status_label.leaveEvent = self.time_header_hover_leave
+        self.time_status_label.setStyleSheet("""
+            QLabel {
+                padding: 2px;
+                border-radius: 8px;
+                background: transparent;
+            }
+        """)
+        time_status_layout.addWidget(self.time_status_label)
+        time_status_layout.addStretch()
+        
+        layout.addLayout(time_status_layout)
+        
+        # Center - Manifest details (carriers) - takes remaining space
+        center_layout = QVBoxLayout()
+        center_layout.setSpacing(0)  # No spacing in center section
+        center_layout.setContentsMargins(0, 0, 0, 0)
+        
+        # Create a horizontal layout for carriers and acknowledgment side by side
+        carriers_ack_layout = QHBoxLayout()
+        carriers_ack_layout.setSpacing(15)  # Reduced spacing between carriers and ack  
+        carriers_ack_layout.setContentsMargins(0, 0, 0, 0)  # No margins
+        
+        # Left part: carriers (will be recreated as clickable labels)
+        self.carriers_widget = QWidget()
+        self.carriers_layout = QVBoxLayout(self.carriers_widget)
+        self.carriers_layout.setContentsMargins(0, 0, 0, 0)
+        self.carriers_layout.setSpacing(1)  # Extremely tight - reduced from 2 to 1
+        carriers_ack_layout.addWidget(self.carriers_widget)
+        
+        # Right part: acknowledgment (consistent fixed width)
+        ack_widget = QWidget()
+        ack_widget.setFixedWidth(250)  # Consistent fixed width for all cards
+        ack_widget.setStyleSheet("background: transparent;")
+        
+        ack_layout = QVBoxLayout(ack_widget)
+        ack_layout.setContentsMargins(0, 0, 0, 0)  # No margins
+        ack_layout.setSpacing(0)  # No spacing between status and details
+        
+        # Acknowledgment status text
+        self.ack_status_label = QLabel("")
+        self.ack_status_label.setFont(QFont("Segoe UI", 18, QFont.Weight.Bold))  # Increased from 16 to 18
+        self.ack_status_label.setStyleSheet("color: #2ed573; background: transparent;")
+        self.ack_status_label.setAlignment(Qt.AlignmentFlag.AlignLeft | Qt.AlignmentFlag.AlignTop)  # Left align
+        ack_layout.addWidget(self.ack_status_label)
+        
+        # Reason/details text
+        self.ack_details_label = QLabel("")
+        self.ack_details_label.setFont(QFont("Segoe UI", 16))  # Increased from 14 to 16
+        self.ack_details_label.setStyleSheet("color: #ffffff; background: transparent;")
+        self.ack_details_label.setWordWrap(True)
+        self.ack_details_label.setAlignment(Qt.AlignmentFlag.AlignLeft | Qt.AlignmentFlag.AlignTop)  # Left align
+        ack_layout.addWidget(self.ack_details_label)
+        
+        ack_layout.addStretch()
+        carriers_ack_layout.addWidget(ack_widget)
+        # Removed addStretch here to let sections naturally space out
+        
+        center_layout.addLayout(carriers_ack_layout)
+        center_layout.addStretch()
+        
+        layout.addLayout(center_layout)
+        
+        # Remove redundant status bar - card border provides sufficient visual indication
+        self.setLayout(layout)
+    
+    def update_styling(self):
+        """Update colors based on status - only borders and font colors on black background"""
+        if self.status == "ACTIVE":
+            border_color = "#ff4757"
+            text_color = "#ffffff"
+            status_color = "#ff4757"
+        elif self.status == "MISSED":
+            border_color = "#c44569"
+            text_color = "#ffffff"
+            status_color = "#c44569"
+        elif self.status == "ACKNOWLEDGED":
+            border_color = "#2ed573"
+            text_color = "#ffffff"
+            status_color = "#2ed573"
+        else:  # OPEN
+            border_color = "#3742fa"
+            text_color = "#ffffff"
+            status_color = "#3742fa"
+        
+        self.setStyleSheet(f"""
+            StatusCard {{
+                background: transparent;
+                border: 2px solid {border_color};
+                border-radius: 12px;
+            }}
+        """)
+        
+        self.time_status_label.setStyleSheet(f"color: {status_color}; background: transparent;")
+    
+    def set_manifests(self, manifests):
+        """Update manifest details"""
+        self.manifests = manifests
+        if manifests:
+            self.update_manifest_display()
+        else:
+            self.details_label.setText("")
+    
+    def set_acknowledgment(self, user_info, reason=None):
+        """Update acknowledgment display with user details"""
+        if user_info:
+            if reason == "late":
+                # For missed deliveries - show user and time
+                self.ack_status_label.setText("Done Late")
+                self.ack_details_label.setText(f"by {user_info}")
+                self.ack_status_label.setStyleSheet("color: #ffb347; background: transparent;")  # Lighter orange
+                self.ack_details_label.setStyleSheet("color: #ffb347; background: transparent;")  # Match color
+            else:
+                # For completed/acknowledged - show user and time
+                self.ack_status_label.setText("Done")
+                self.ack_details_label.setText(f"by {user_info}")
+                self.ack_status_label.setStyleSheet("color: #2ed573; background: transparent;")  # Green for done
+                self.ack_details_label.setStyleSheet("color: #2ed573; background: transparent;")  # Match color
+        else:
+            # Clear acknowledgment
+            self.ack_status_label.setText("")
+            self.ack_details_label.setText("")
+    
+    def card_clicked(self, event):
+        """Handle card click - disabled for now, use double-click on time header"""
+        pass  # Disabled card clicking, only time header double-click works
+    
+    def card_hover_enter(self, event):
+        """Handle mouse enter on card border"""
+        # Add subtle card border highlight
+        current_style = self.styleSheet()
+        if "border: 3px solid" not in current_style:
+            # Increase border thickness slightly for hover effect
+            for status in ["ACTIVE", "MISSED", "ACKNOWLEDGED", "OPEN"]:
+                if status.lower() in current_style.lower():
+                    current_style = current_style.replace("border: 2px solid", "border: 3px solid")
+                    break
+            self.setStyleSheet(current_style)
+    
+    def card_hover_leave(self, event):
+        """Handle mouse leave on card"""
+        # Remove card border highlight
+        current_style = self.styleSheet()
+        current_style = current_style.replace("border: 3px solid", "border: 2px solid")
+        self.setStyleSheet(current_style)
+        # Also ensure time header highlight is removed
+        self.time_header_hover_leave(event)
+    
+    def time_header_double_clicked(self, event):
+        """Handle double-click on time header to acknowledge whole card"""
+        if (self.status in ["ACTIVE", "MISSED"] and 
+            self.parent_display and 
+            hasattr(self.parent_display, 'acknowledge_time_slot')):
+            self.parent_display.acknowledge_time_slot(self.time_str)
+    
+    def time_header_hover_enter(self, event):
+        """Handle mouse enter on time header - no background colors"""
+        # No background color hover effects - keep clean appearance
+        pass
+    
+    def time_header_hover_leave(self, event):
+        """Handle mouse leave on time header - no background colors"""
+        # No background color hover effects - keep clean appearance
+        pass
+    
+    def update_manifest_display(self):
+        """Update the manifest display with individual clickable carriers"""
+        # Clear existing carrier labels
+        for i in reversed(range(self.carriers_layout.count())):
+            child = self.carriers_layout.itemAt(i).widget()
+            if child:
+                child.setParent(None)
+        
+        if self.manifests:
+            for carrier, status in self.manifests:
+                # Create individual carrier label
+                carrier_label = QLabel()
+                carrier_label.setFont(QFont("Segoe UI", 18))  # Increased from 16 to 18
+                carrier_label.setAlignment(Qt.AlignmentFlag.AlignLeft | Qt.AlignmentFlag.AlignTop)
+                # No padding at all for minimal line height
+                
+                # Set text and styling based on status (clean display)
+                if status == "Acknowledged":
+                    carrier_label.setText(f"{carrier} [DONE]")
+                    carrier_label.setStyleSheet("color: #2ed573; background: transparent;")
+                elif status == "AcknowledgedLate":
+                    carrier_label.setText(f"{carrier} [DONE LATE]")
+                    carrier_label.setStyleSheet("color: #ffb347; background: transparent;")
+                elif status == "Active":
+                    carrier_label.setText(carrier)
+                    carrier_label.setStyleSheet("color: #ffffff; background: transparent; border-radius: 5px;")
+                    # Make active items clickable and hoverable
+                    carrier_label.mousePressEvent = lambda event, c=carrier: self.acknowledge_single_carrier(c)
+                    carrier_label.enterEvent = lambda event, label=carrier_label: self.carrier_hover_enter(label, "#ff4757")
+                    carrier_label.leaveEvent = lambda event, label=carrier_label: self.carrier_hover_leave(label)
+                elif status == "Missed":
+                    carrier_label.setText(f"{carrier} [MISSED]")
+                    carrier_label.setStyleSheet("color: #ff4757; background: transparent; border-radius: 5px;")
+                    # Make missed items clickable and hoverable
+                    carrier_label.mousePressEvent = lambda event, c=carrier: self.acknowledge_single_carrier(c)
+                    carrier_label.enterEvent = lambda event, label=carrier_label: self.carrier_hover_enter(label, "#c44569")
+                    carrier_label.leaveEvent = lambda event, label=carrier_label: self.carrier_hover_leave(label)
+                else:  # Open
+                    carrier_label.setText(carrier)
+                    carrier_label.setStyleSheet("color: #ffffff; background: transparent;")
+                
+                self.carriers_layout.addWidget(carrier_label)
+            
+            # Extremely compact height calculation for minimal vertical space
+            line_count = len(self.manifests)
+            base_height = 35       # Even smaller base height
+            line_height = 22       # Tight but readable height per carrier line  
+            padding = 10           # Minimal padding
+            new_height = base_height + (line_count * line_height) + padding
+            self.setFixedHeight(new_height)  # Use fixed height for consistent appearance
+            
+            # Update overall card status
+            self.update_card_status()
+    
+    def carrier_hover_enter(self, label, color):
+        """Handle hover enter on individual carrier - no background colors"""
+        # Just add a subtle text glow effect by increasing brightness
+        current_style = label.styleSheet()
+        # For now, just leave the existing style - no background colors
+        pass
+    
+    def carrier_hover_leave(self, label):
+        """Handle hover leave on individual carrier - no background colors"""
+        # No background colors to remove, just leave as is
+        pass
+    
+    def hex_to_rgba(self, hex_color, alpha):
+        """Convert hex color to RGBA string"""
+        hex_color = hex_color.lstrip('#')
+        r = int(hex_color[0:2], 16)
+        g = int(hex_color[2:4], 16)
+        b = int(hex_color[4:6], 16)
+        return f"({r}, {g}, {b}, {alpha})"
+    
+    def acknowledge_single_carrier(self, carrier):
+        """Acknowledge a single carrier"""
+        if self.parent_display and hasattr(self.parent_display, 'acknowledge_single_carrier'):
+            self.parent_display.acknowledge_single_carrier(self.time_str, carrier)
+    
+    def update_card_status(self):
+        """Update card status based on manifest states"""
+        if not self.manifests:
+            return
+            
+        acknowledged_count = sum(1 for _, status in self.manifests if status in ["Acknowledged", "AcknowledgedLate"])
+        missed_count = sum(1 for _, status in self.manifests if status == "Missed")
+        active_count = sum(1 for _, status in self.manifests if status == "Active")
+        total_count = len(self.manifests)
+        
+        # Determine card status and acknowledgment display
+        if acknowledged_count == total_count:
+            # All acknowledged
+            self.status = "ACKNOWLEDGED"
+            # Show acknowledgment info from the first acknowledged item
+            if self.parent_display:
+                ack_info = self.get_acknowledgment_info()
+                if ack_info:
+                    user_name = ack_info.get('user', 'Unknown')
+                    reason = ack_info.get('reason', '')
+                    timestamp = ack_info.get('timestamp', '')
+                    if timestamp:
+                        try:
+                            # Parse timestamp and format time
+                            ack_time = datetime.fromisoformat(timestamp)
+                            time_str = ack_time.strftime('%H:%M')
+                            if reason:  # Late acknowledgment
+                                self.set_acknowledgment(f"{user_name} at {time_str}", "late")
+                            else:  # Regular acknowledgment
+                                self.set_acknowledgment(f"{user_name} at {time_str}")
+                        except:
+                            if reason:
+                                self.set_acknowledgment(user_name, "late")
+                            else:
+                                self.set_acknowledgment(user_name)
+                    else:
+                        if reason:
+                            self.set_acknowledgment(user_name, "late")
+                        else:
+                            self.set_acknowledgment(user_name)
+        elif missed_count > 0:
+            # Has missed items - card becomes MISSED
+            self.status = "MISSED"
+            if acknowledged_count > 0:
+                # Partially acknowledged
+                self.ack_status_label.setText(f"{acknowledged_count}/{total_count} Done")
+                self.ack_status_label.setStyleSheet("color: #ffb347; background: transparent;")  # Lighter orange for partial
+                self.ack_details_label.setText("")
+            else:
+                self.set_acknowledgment("")  # Clear acknowledgment display
+        elif active_count > 0:
+            # Has active items - card becomes ACTIVE
+            self.status = "ACTIVE"
+            if acknowledged_count > 0:
+                # Partially acknowledged
+                self.ack_status_label.setText(f"{acknowledged_count}/{total_count} Done")
+                self.ack_status_label.setStyleSheet("color: #ffb347; background: transparent;")  # Lighter orange for partial
+                self.ack_details_label.setText("")
+            else:
+                self.set_acknowledgment("")  # Clear acknowledgment display
+        else:
+            # Open items
+            self.status = "OPEN"
+            self.set_acknowledgment("")  # Clear acknowledgment display
+        
+        # Update the combined time/status label - change ACKNOWLEDGED to DONE
+        display_status = "DONE" if self.status == "ACKNOWLEDGED" else self.status
+        self.time_status_label.setText(f"{self.time_str} - {display_status}")
+        self.update_styling()
+    
+    def get_acknowledgment_info(self):
+        """Get acknowledgment info for the first acknowledged item"""
+        if not self.parent_display:
+            return None
+            
+        try:
+            acks = self.parent_display.load_acknowledgments()
+            today = datetime.now().date().isoformat()
+            
+            for carrier, status in self.manifests:
+                if status in ["Acknowledged", "AcknowledgedLate"]:
+                    ack_key = f"{today}_{self.time_str}_{carrier}"
+                    if ack_key in acks:
+                        return acks[ack_key]
+            return None
+        except:
+            return None
+
+
+class AlertDisplay(QWidget):
+    def __init__(self):
+        super().__init__()
+        self.setWindowTitle("Manifest Times")
+        self.setMinimumSize(1200, 800)
+        
+        # Alert state management
+        self.alert_active = False
+        
+        # Set window icon
+        icon_path = os.path.join(os.path.dirname(os.path.abspath(__file__)), 'resources', 'icon.ico')
+        if os.path.exists(icon_path):
+            window_icon = QIcon(icon_path)
+            if not window_icon.isNull():
+                self.setWindowIcon(window_icon)
+        
+        self.status_cards = {}
+        self.clock_timer = None
+        self.refresh_timer = None
+        self.setup_ui()
+        self.setup_timers()
+        self.apply_dark_theme()
+        self.populate_data()
+    
+    def setup_ui(self):
+        """Create the modern mission control layout"""
+        main_layout = QVBoxLayout()
+        main_layout.setContentsMargins(30, 10, 30, 30)  # Reduced top margin from 30 to 10
+        main_layout.setSpacing(10)  # Reduced spacing from 20 to 10
+        
+        # Header section
+        header_layout = QHBoxLayout()
+        
+        # Title
+        title_label = QLabel("MANIFEST TIMES")
+        title_label.setFont(QFont("Segoe UI", 36, QFont.Weight.Bold))
+        title_label.setStyleSheet("color: #ffffff; padding: 0px;")  # Removed bottom padding
+        header_layout.addWidget(title_label)
+        
+        header_layout.addStretch()
+        
+        # Settings button with cog icon
+        self.settings_btn = QPushButton("⚙️")
+        self.settings_btn.setFont(QFont("Segoe UI", 16, QFont.Weight.Bold))
+        self.settings_btn.setFixedSize(60, 40)
+        self.settings_btn.setStyleSheet("""
+            QPushButton {
+                background-color: #2c2c54;
+                color: #ffffff;
+                border: 2px solid #3742fa;
+                border-radius: 20px;
+                padding: 0px;
+            }
+            QPushButton:hover {
+                background-color: #3742fa;
+            }
+            QPushButton:pressed {
+                background-color: #1f2ecc;
+            }
+        """)
+        self.settings_btn.clicked.connect(self.show_settings_dialog)
+        header_layout.addWidget(self.settings_btn)
+        
+        # Refresh Data button in header with refresh icon
+        self.reload_btn = QPushButton("🔄")
+        self.reload_btn.setFont(QFont("Segoe UI", 16, QFont.Weight.Bold))
+        self.reload_btn.setFixedSize(60, 40)
+        self.reload_btn.setStyleSheet("""
+            QPushButton {
+                background-color: #2c2c54;
+                color: #ffffff;
+                border: 2px solid #3742fa;
+                border-radius: 20px;
+                padding: 8px 16px;
+            }
+            QPushButton:hover {
+                background-color: #3742fa;
+            }
+            QPushButton:pressed {
+                background-color: #1f2ecc;
+            }
+        """)
+        self.reload_btn.clicked.connect(self.populate_data)
+        header_layout.addWidget(self.reload_btn)
+        
+        # Clock
+        self.clock_label = QLabel()
+        self.clock_label.setFont(QFont("Segoe UI", 32, QFont.Weight.Bold))
+        self.clock_label.setStyleSheet("color: #3742fa; padding: 0px; margin-left: 20px;")  # Removed bottom padding
+        header_layout.addWidget(self.clock_label)
+        
+        main_layout.addLayout(header_layout)
+        
+        # Status summary bar
+        self.summary_label = QLabel("SYSTEM NOMINAL")
+        self.summary_label.setFont(QFont("Segoe UI", 18, QFont.Weight.Bold))
+        self.summary_label.setAlignment(Qt.AlignmentFlag.AlignCenter)
+        self.summary_label.setStyleSheet("""
+            background-color: #2ed573;
+            color: #000000;
+            padding: 12px;
+            border-radius: 8px;
+            margin-bottom: 10px;
+        """)
+        main_layout.addWidget(self.summary_label)
+        
+        # Scrollable cards area
+        scroll_area = QScrollArea()
+        scroll_area.setWidgetResizable(True)
+        scroll_area.setHorizontalScrollBarPolicy(Qt.ScrollBarPolicy.ScrollBarAlwaysOff)
+        scroll_area.setVerticalScrollBarPolicy(Qt.ScrollBarPolicy.ScrollBarAsNeeded)
+        scroll_area.setStyleSheet("""
+            QScrollArea {
+                border: none;
+                background-color: transparent;
+            }
+            QScrollBar:vertical {
+                background-color: #2c2c54;
+                width: 12px;
+                border-radius: 6px;
+            }
+            QScrollBar::handle:vertical {
+                background-color: #3742fa;
+                border-radius: 6px;
+            }
+        """)
+        
+        # Cards container
+        cards_widget = QWidget()
+        self.cards_layout = QGridLayout(cards_widget)
+        self.cards_layout.setSpacing(15)  # Reduced from 20 to 15
+        self.cards_layout.setContentsMargins(10, 10, 10, 10)
+        
+        scroll_area.setWidget(cards_widget)
+        main_layout.addWidget(scroll_area)
+        self.setLayout(main_layout)
+    
+    def apply_dark_theme(self):
+        """Apply SpaceX-style dark theme"""
+        self.setStyleSheet("""
+            QWidget {
+                background-color: #0f0f23;
+                color: #ffffff;
+                font-family: 'Segoe UI', Arial, sans-serif;
+            }
+            QLabel {
+                background: transparent;
+                color: inherit;
+            }
+            QPushButton {
+                background-color: #3742fa;
+                color: #ffffff;
+                border: none;
+                border-radius: 8px;
+                padding: 15px 30px;
+                font-weight: bold;
+            }
+            QPushButton:hover {
+                background-color: #4f69ff;
+            }
+            QPushButton:pressed {
+                background-color: #2c35e6;
+            }
+            QMessageBox {
+                background-color: #1a1a2e;
+                color: #ffffff;
+                font-size: 16px;
+            }
+            QMessageBox QLabel {
+                color: #ffffff;
+                font-size: 16px;
+            }
+            QMessageBox QPushButton {
+                background-color: #3742fa;
+                color: #ffffff;
+                border: none;
+                border-radius: 5px;
+                padding: 8px 16px;
+                font-size: 14px;
+                font-weight: bold;
+                min-width: 80px;
+            }
+            QMessageBox QPushButton:hover {
+                background-color: #4f69ff;
+            }
+        """)
+    
+    def setup_timers(self):
+        """Setup update timers"""
+        # Stop existing timers if they exist
+        if hasattr(self, 'clock_timer') and self.clock_timer:
+            self.clock_timer.stop()
+        if hasattr(self, 'refresh_timer') and self.refresh_timer:
+            self.refresh_timer.stop()
+        
+        # Clock timer
+        self.clock_timer = QTimer(self)
+        self.clock_timer.timeout.connect(self.update_clock)
+        self.clock_timer.start(1000)
+        
+        # Data refresh timer
+        self.refresh_timer = QTimer(self)
+        self.refresh_timer.timeout.connect(self.populate_data)
+        self.refresh_timer.start(30000)  # 30 seconds
+        
+        self.update_clock()
+    
+    def update_clock(self):
+        """Update the clock display"""
+        now = datetime.now()
+        self.clock_label.setText(now.strftime('%H:%M:%S'))
+    
+    def load_config(self):
+        """Load configuration from settings-specified folder or fallback"""
+        try:
+            settings = self.load_settings()
+            data_folder = settings.get('data_folder', '')
+            
+            # Try settings-specified folder first
+            if data_folder and os.path.exists(data_folder):
+                config_path = os.path.join(data_folder, 'config.json')
+                if os.path.exists(config_path):
+                    with open(config_path, 'r', encoding='utf-8') as f:
+                        return json.load(f)
+            
+            # Fallback to default locations
+            for folder in ['data', 'app_data', '.']:
+                config_path = os.path.join(os.path.dirname(__file__), folder, 'config.json')
+                if os.path.exists(config_path):
+                    with open(config_path, 'r', encoding='utf-8') as f:
+                        return json.load(f)
+            
+            # Return fallback data if no config found
+            return load_config()
+            
+        except Exception:
+            return load_config()
+    
+    def populate_data(self):
+        """Populate cards with manifest data"""
+        # Clear existing cards
+        for card in self.status_cards.values():
+            card.setParent(None)
+        self.status_cards.clear()
+        
+        # Load configuration
+        config = self.load_config()
+        manifests = config.get('manifests', [])
+        
+        if not manifests:
+            # Show "no data" message
+            no_data_label = QLabel("NO MANIFEST DATA AVAILABLE")
+            no_data_label.setFont(QFont("Segoe UI", 24, QFont.Weight.Bold))
+            no_data_label.setAlignment(Qt.AlignmentFlag.AlignCenter)
+            no_data_label.setStyleSheet("color: #ff4757; padding: 100px;")
+            self.cards_layout.addWidget(no_data_label, 0, 0)
+            self.update_summary("NO DATA")
+            return
+        
+        # Sort manifests by time
+        manifests = sorted(manifests, key=lambda m: m['time'])
+        
+        # Load acknowledgments
+        acks = self.load_acknowledgments()
+        
+        now = datetime.now()
+        today = now.date().isoformat()
+        
+        # Create status cards - one per row
+        row = 0
+        cols_per_row = 1  # One card per row for wider layout
+        
+        active_count = 0
+        missed_count = 0
+        open_count = 0
+        acked_count = 0
+        
+        for manifest in manifests:
+            time_str = manifest['time']
+            
+            # Create status card with parent reference
+            card = StatusCard(time_str, parent_display=self)
+            self.status_cards[time_str] = card
+            
+            # Process carriers for this time - determine overall time slot status first
+            manifest_data = []
+            time_slot_status = get_manifest_status(time_str, now)
+            
+            for carrier in manifest.get('carriers', []):
+                ack_key = f"{today}_{time_str}_{carrier}"
+                is_acked = ack_key in acks
+                
+                if is_acked:
+                    # Check if it was a late acknowledgment
+                    ack_info = acks[ack_key]
+                    reason = ack_info.get('reason', '')
+                    if reason and reason.strip():  # Has reason = late acknowledgment
+                        status = "AcknowledgedLate"
+                    else:
+                        status = "Acknowledged"
+                    acked_count += 1
+                    
+                    # Set acknowledgment display
+                    user_name = ack_info.get('user', 'Unknown')
+                    card.set_acknowledgment(user_name, reason if reason else None)
+                else:
+                    # Use the time slot status for all non-acknowledged items
+                    status = time_slot_status
+                    if status == "Active":
+                        active_count += 1
+                    elif status == "Missed":
+                        missed_count += 1
+                    else:
+                        open_count += 1
+                
+                manifest_data.append((carrier, status))
+            
+            card.set_manifests(manifest_data)
+            
+            # Add to grid - one per row
+            self.cards_layout.addWidget(card, row, 0)
+            row += 1
+        
+        # Update alert state
+        self.alert_active = (active_count > 0 or missed_count > 0)
+        
+        # Update summary with next manifest countdown
+        next_manifest_info = self.get_next_manifest_info(manifests, now)
+        if active_count > 0:
+            self.update_summary("ACTIVE ALERTS", "#ff4757")
+        elif missed_count > 0:
+            self.update_summary("MISSED MANIFESTS", "#c44569")
+        elif next_manifest_info:
+            self.update_summary(next_manifest_info, "#3742fa")
+        else:
+            self.update_summary("ALL SYSTEMS NOMINAL", "#2ed573")
+    
+    def get_next_manifest_info(self, manifests, now):
+        """Get countdown to next manifest"""
+        try:
+            current_time = now.time()
+            next_manifest = None
+            
+            for manifest in manifests:
+                manifest_time = datetime.strptime(manifest['time'], '%H:%M').time()
+                
+                # Check if this manifest is in the future today
+                if manifest_time > current_time:
+                    next_manifest = manifest
+                    break
+            
+            # If no manifest found for today, check tomorrow's first manifest
+            if not next_manifest and manifests:
+                next_manifest = manifests[0]
+                # Calculate time to tomorrow's first manifest
+                tomorrow = now.replace(hour=0, minute=0, second=0, microsecond=0) + timedelta(days=1)
+                manifest_time = datetime.strptime(next_manifest['time'], '%H:%M').time()
+                next_time = tomorrow.replace(hour=manifest_time.hour, minute=manifest_time.minute)
+            else:
+                # Calculate time to today's next manifest
+                manifest_time = datetime.strptime(next_manifest['time'], '%H:%M').time()
+                next_time = now.replace(hour=manifest_time.hour, minute=manifest_time.minute, second=0, microsecond=0)
+            
+            # Calculate time difference
+            time_diff = next_time - now
+            
+            # Convert to hours and minutes
+            total_seconds = int(time_diff.total_seconds())
+            hours = total_seconds // 3600
+            minutes = (total_seconds % 3600) // 60
+            
+            if hours > 0:
+                return f"NEXT MANIFEST IN {hours}h {minutes}m ({next_manifest['time']})"
+            else:
+                return f"NEXT MANIFEST IN {minutes}m ({next_manifest['time']})"
+                
+        except Exception:
+            return "NEXT MANIFEST OPEN"
+    
+    def update_summary(self, text, color="#2ed573"):
+        """Update the status summary bar"""
+        text_color = "#000000" if color == "#2ed573" else "#ffffff"
+        self.summary_label.setText(text)
+        self.summary_label.setStyleSheet(f"""
+            background-color: {color};
+            color: {text_color};
+            padding: 12px;
+            border-radius: 8px;
+            margin-bottom: 10px;
+        """)
+    
+    def load_acknowledgments(self):
+        """Load acknowledgment data from ack.json"""
+        try:
+            settings = self.load_settings()
+            data_folder = settings.get('data_folder', '')
+            
+            # Try settings-specified folder first
+            if data_folder and os.path.exists(data_folder):
+                ack_path = os.path.join(data_folder, 'ack.json')
+            else:
+                # Fallback to local ack.json
+                ack_path = os.path.join(os.path.dirname(__file__), 'ack.json')
+            
+            if not os.path.exists(ack_path):
+                return {}
+            
+            with open(ack_path, 'r', encoding='utf-8') as f:
+                ack_data = json.load(f)
+            
+            # Convert to lookup dict
+            acks = {}
+            today = datetime.now().date().isoformat()
+            
+            for ack in ack_data:
+                if ack.get('date') == today:
+                    key = f"{ack['date']}_{ack['manifest_time']}_{ack['carrier']}"
+                    acks[key] = ack
+            
+            return acks
+            
+        except Exception as e:
+            return {}
+    
+    def show_settings_dialog(self):
+        """Show settings configuration dialog"""
+        dialog = QDialog(self)
+        dialog.setWindowTitle("Settings")
+        dialog.setFixedSize(500, 300)
+        dialog.setStyleSheet("""
+            QDialog {
+                background-color: #1a1a2e;
+                color: #ffffff;
+            }
+            QLabel {
+                color: #ffffff;
+                font-size: 14px;
+            }
+            QLineEdit {
+                background-color: #2c2c54;
+                border: 2px solid #3742fa;
+                border-radius: 5px;
+                padding: 8px;
+                color: #ffffff;
+                font-size: 14px;
+            }
+            QLineEdit:focus {
+                border-color: #4f69ff;
+            }
+            QLineEdit.error {
+                border-color: #ff4757;
+                background-color: #3d1a1a;
+            }
+            QPushButton {
+                background-color: #3742fa;
+                color: #ffffff;
+                border: none;
+                border-radius: 5px;
+                padding: 8px 16px;
+                font-weight: bold;
+            }
+            QPushButton:hover {
+                background-color: #4f69ff;
+            }
+            QPushButton:disabled {
+                background-color: #555555;
+                color: #999999;
+            }
+            .status-label {
+                font-size: 12px;
+                padding: 5px;
+                border-radius: 3px;
+            }
+            .status-valid {
+                color: #2ed573;
+                background-color: #1b2d1b;
+            }
+            .status-invalid {
+                color: #ff4757;
+                background-color: #3d1a1a;
+            }
+        """)
+        
+        layout = QVBoxLayout()
+        form_layout = QFormLayout()
+        
+        # Load current settings - ensure we get the actual values
+        current_settings = self.load_settings()
+        
+        # Username field
+        self.username_edit = QLineEdit()
+        username_value = current_settings.get('username', '')
+        self.username_edit.setText(username_value)
+        self.username_edit.setPlaceholderText("Enter username...")
+        form_layout.addRow("Username:", self.username_edit)
+        
+        # Folder path field with validation
+        folder_widget = QWidget()
+        folder_layout = QVBoxLayout(folder_widget)
+        folder_layout.setContentsMargins(0, 0, 0, 0)
+        folder_layout.setSpacing(5)
+        
+        self.folder_edit = QLineEdit()
+        folder_value = current_settings.get('data_folder', '')
+        self.folder_edit.setText(folder_value)
+        self.folder_edit.setPlaceholderText("Enter folder path for JSON files...")
+        self.folder_edit.textChanged.connect(self.validate_folder_path)
+        
+        # Status label for folder validation
+        self.folder_status_label = QLabel("")
+        self.folder_status_label.setWordWrap(True)
+        
+        folder_layout.addWidget(self.folder_edit)
+        folder_layout.addWidget(self.folder_status_label)
+        
+        form_layout.addRow("Data Folder:", folder_widget)
+        layout.addLayout(form_layout)
+        
+        # Buttons
+        self.button_box = QDialogButtonBox(QDialogButtonBox.StandardButton.Ok | QDialogButtonBox.StandardButton.Cancel)
+        self.button_box.accepted.connect(lambda: self.save_settings_and_close(dialog, current_settings))
+        self.button_box.rejected.connect(dialog.reject)
+        layout.addWidget(self.button_box)
+        
+        dialog.setLayout(layout)
+        
+        # Initial validation
+        self.validate_folder_path()
+        
+        dialog.exec()
+    
+    def validate_folder_path(self):
+        """Validate the folder path in real-time"""
+        folder_path = self.folder_edit.text().strip()
+        
+        if not folder_path:
+            # Empty path is allowed (will use default)
+            self.folder_status_label.setText("Using default data locations")
+            self.folder_status_label.setProperty("class", "status-label")
+            self.folder_status_label.setStyleSheet("color: #999999; font-size: 12px;")
+            self.folder_edit.setProperty("class", "")
+            self.folder_edit.setStyleSheet("")
+            self.button_box.button(QDialogButtonBox.StandardButton.Ok).setEnabled(True)
+            return True
+        
+        # Check if path exists and is accessible
+        if os.path.exists(folder_path):
+            if os.path.isdir(folder_path):
+                # Check if we can write to this directory
+                try:
+                    test_file = os.path.join(folder_path, 'test_write.tmp')
+                    with open(test_file, 'w') as f:
+                        f.write('test')
+                    os.remove(test_file)
+                    
+                    self.folder_status_label.setText("✓ Valid folder with write access")
+                    self.folder_status_label.setProperty("class", "status-label status-valid")
+                    self.folder_status_label.setStyleSheet("color: #2ed573; background-color: #1b2d1b; font-size: 12px; padding: 5px; border-radius: 3px;")
+                    self.folder_edit.setProperty("class", "")
+                    self.folder_edit.setStyleSheet("")
+                    self.button_box.button(QDialogButtonBox.StandardButton.Ok).setEnabled(True)
+                    return True
+                except (PermissionError, OSError):
+                    self.folder_status_label.setText("✗ Folder exists but no write permission")
+                    self.folder_status_label.setProperty("class", "status-label status-invalid")
+                    self.folder_status_label.setStyleSheet("color: #ff4757; background-color: #3d1a1a; font-size: 12px; padding: 5px; border-radius: 3px;")
+                    self.folder_edit.setProperty("class", "error")
+                    self.folder_edit.setStyleSheet("border-color: #ff4757; background-color: #3d1a1a;")
+                    self.button_box.button(QDialogButtonBox.StandardButton.Ok).setEnabled(False)
+                    return False
+            else:
+                self.folder_status_label.setText("✗ Path exists but is not a folder")
+                self.folder_status_label.setProperty("class", "status-label status-invalid")
+                self.folder_status_label.setStyleSheet("color: #ff4757; background-color: #3d1a1a; font-size: 12px; padding: 5px; border-radius: 3px;")
+                self.folder_edit.setProperty("class", "error")
+                self.folder_edit.setStyleSheet("border-color: #ff4757; background-color: #3d1a1a;")
+                self.button_box.button(QDialogButtonBox.StandardButton.Ok).setEnabled(False)
+                return False
+        else:
+            self.folder_status_label.setText("✗ Folder does not exist")
+            self.folder_status_label.setProperty("class", "status-label status-invalid")
+            self.folder_status_label.setStyleSheet("color: #ff4757; background-color: #3d1a1a; font-size: 12px; padding: 5px; border-radius: 3px;")
+            self.folder_edit.setProperty("class", "error")
+            self.folder_edit.setStyleSheet("border-color: #ff4757; background-color: #3d1a1a;")
+            self.button_box.button(QDialogButtonBox.StandardButton.Ok).setEnabled(False)
+            return False
+    
+    def load_settings(self):
+        """Load settings from settings.json"""
+        try:
+            settings_path = os.path.join(os.path.dirname(__file__), 'settings.json')
+            if os.path.exists(settings_path):
+                with open(settings_path, 'r', encoding='utf-8') as f:
+                    settings = json.load(f)
+                    # Ensure we have default values for missing keys
+                    return {
+                        'username': settings.get('username', ''),
+                        'data_folder': settings.get('data_folder', '')
+                    }
+        except Exception:
+            pass
+        return {'username': '', 'data_folder': ''}
+    
+    def save_settings_and_close(self, dialog, original_settings):
+        """Save settings with validation and preserve existing values"""
+        try:
+            # Get new values
+            new_username = self.username_edit.text().strip()
+            new_folder = self.folder_edit.text().strip()
+            
+            # Validate folder if provided
+            if new_folder and not self.validate_folder_path():
+                QMessageBox.warning(self, "Invalid Folder", 
+                                   "Please fix the folder path before saving.")
+                return
+            
+            # Preserve existing values if new ones are empty (except for intentional clearing)
+            final_username = new_username if new_username else original_settings.get('username', '')
+            final_folder = new_folder  # Allow empty folder (uses defaults)
+            
+            settings = {
+                'username': final_username,
+                'data_folder': final_folder
+            }
+            
+            settings_path = os.path.join(os.path.dirname(__file__), 'settings.json')
+            with open(settings_path, 'w', encoding='utf-8') as f:
+                json.dump(settings, f, indent=2)
+            
+            QMessageBox.information(self, "Settings", "Settings saved successfully!")
+            dialog.accept()
+            
+        except Exception as e:
+            QMessageBox.warning(self, "Error", f"Failed to save settings: {e}")
+    
+    def closeEvent(self, event):
+        """Handle window close event with proper cleanup"""
+        # Stop all timers
+        if self.clock_timer:
+            self.clock_timer.stop()
+            self.clock_timer = None
+        if self.refresh_timer:
+            self.refresh_timer.stop()
+            self.refresh_timer = None
+        
+        # Clean up status cards
+        for card in self.status_cards.values():
+            card.setParent(None)
+        self.status_cards.clear()
+        
+        # Accept the close event
+        event.accept()
+    
+    def acknowledge_time_slot(self, time_str):
+        """Acknowledge all items in a time slot"""
+        if time_str not in self.status_cards:
+            return
+            
+        card = self.status_cards[time_str]
+        if not card.manifests:
+            return
+        
+        # Get username from settings
+        settings = self.load_settings()
+        username = settings.get('username', '').strip()
+        
+        if not username:
+            QMessageBox.warning(self, "No Username", 
+                              "Please set your username in Settings before acknowledging.")
+            return
+        
+        # Determine if this is a late acknowledgment (any missed items)
+        has_missed = any(status == "Missed" for _, status in card.manifests)
+        
+        # Save acknowledgments immediately
+        try:
+            current_time = datetime.now()
+            today = current_time.date().isoformat()
+            timestamp = current_time.isoformat()
+            
+            # Load existing acknowledgments
+            settings = self.load_settings()
+            data_folder = settings.get('data_folder', '')
+            
+            if data_folder and os.path.exists(data_folder):
+                ack_path = os.path.join(data_folder, 'ack.json')
+            else:
+                ack_path = os.path.join(os.path.dirname(__file__), 'ack.json')
+            
+            # Load existing data
+            ack_data = []
+            if os.path.exists(ack_path):
+                try:
+                    with open(ack_path, 'r', encoding='utf-8') as f:
+                        ack_data = json.load(f)
+                except:
+                    ack_data = []
+            
+            # Add acknowledgments for all carriers in this time slot
+            for carrier, status in card.manifests:
+                # Check if already acknowledged
+                existing_ack = None
+                for ack in ack_data:
+                    if (ack.get('date') == today and 
+                        ack.get('manifest_time') == time_str and 
+                        ack.get('carrier') == carrier):
+                        existing_ack = ack
+                        break
+                
+                if not existing_ack:
+                    # Add new acknowledgment
+                    ack_entry = {
+                        'date': today,
+                        'manifest_time': time_str,
+                        'carrier': carrier,
+                        'user': username,
+                        'reason': 'Done Late' if has_missed else '',
+                        'timestamp': timestamp
+                    }
+                    ack_data.append(ack_entry)
+            
+            # Save to file
+            with open(ack_path, 'w', encoding='utf-8') as f:
+                json.dump(ack_data, f, indent=2)
+            
+            # Refresh display immediately to show changes
+            self.populate_data()
+            
+            # No success dialog - visual feedback from card color change is sufficient
+            
+        except Exception as e:
+            QMessageBox.warning(self, "Error", f"Failed to save acknowledgment: {e}")
+    
+    def acknowledge_single_carrier(self, time_str, carrier):
+        """Acknowledge a single carrier for a specific time"""
+        # Get username from settings
+        settings = self.load_settings()
+        username = settings.get('username', '').strip()
+        
+        if not username:
+            QMessageBox.warning(self, "No Username", 
+                              "Please set your username in Settings before acknowledging.")
+            return
+        
+        # Determine if this is a late acknowledgment based on card status
+        is_late = False
+        if time_str in self.status_cards:
+            card = self.status_cards[time_str]
+            # Check if this specific carrier is in missed status
+            for car, status in card.manifests:
+                if car == carrier and status == "Missed":
+                    is_late = True
+                    break
+        
+        try:
+            current_time = datetime.now()
+            today = current_time.date().isoformat()
+            timestamp = current_time.isoformat()
+            
+            # Load existing acknowledgments
+            settings = self.load_settings()
+            data_folder = settings.get('data_folder', '')
+            
+            if data_folder and os.path.exists(data_folder):
+                ack_path = os.path.join(data_folder, 'ack.json')
+            else:
+                ack_path = os.path.join(os.path.dirname(__file__), 'ack.json')
+            
+            # Load existing data
+            ack_data = []
+            if os.path.exists(ack_path):
+                try:
+                    with open(ack_path, 'r', encoding='utf-8') as f:
+                        ack_data = json.load(f)
+                except:
+                    ack_data = []
+            
+            # Check if already acknowledged
+            existing_ack = None
+            for ack in ack_data:
+                if (ack.get('date') == today and 
+                    ack.get('manifest_time') == time_str and 
+                    ack.get('carrier') == carrier):
+                    existing_ack = ack
+                    break
+            
+            if not existing_ack:
+                # Add new acknowledgment
+                ack_entry = {
+                    'date': today,
+                    'manifest_time': time_str,
+                    'carrier': carrier,
+                    'user': username,
+                    'reason': 'Done Late' if is_late else '',
+                    'timestamp': timestamp
+                }
+                ack_data.append(ack_entry)
+                
+                # Save to file
+                with open(ack_path, 'w', encoding='utf-8') as f:
+                    json.dump(ack_data, f, indent=2)
+                
+                # Refresh display immediately to show changes
+                self.populate_data()
+            
+        except Exception as e:
+            QMessageBox.warning(self, "Error", f"Failed to save acknowledgment: {e}")
+
+
+
+if __name__ == "__main__":
+    app = QApplication(sys.argv)
+    window = AlertDisplay()
+    window.show()
+    sys.exit(app.exec())
