+++ v2.md
@@ -1,0 +1,71 @@
+# Manifest Alerts V2 Rebuild Log
+
+## Problem Summary
+- Original issue: Window movement during acknowledgments in fullscreen/maximized modes
+- Agent made multiple failed attempts to fix window state preservation
+- Code became increasingly complex and broken
+- Acknowledgment functionality completely non-functional
+- Decision: Complete rebuild starting with clean tree UI
+
+## V2 Rebuild Plan
+1. **Strip back to basics**: Remove all snooze, complex window management, flashing
+2. **Core tree functionality**: Simple manifest display with acknowledgment
+3. **Clean acknowledgment flow**: No window movement, just data updates
+4. **Add features incrementally**: Build up functionality piece by piece
+
+## Rebuild Progress
+
+### Step 1: Clean Tree UI (COMPLETED)
+- ✅ Remove snooze system
+- ✅ Remove flashing/background effects  
+- ✅ Remove complex window state management
+- ✅ Simple tree with manifest display
+- ✅ Basic acknowledgment that just logs without UI disruption
+
+**Changes Made:**
+- Complete rewrite of alert_display.py (200 lines vs 1400+ before)
+- Simple tree structure with time groups and carriers
+- Color coding: Blue=Pending, Red=Active, Dark Red=Missed, Green=Acknowledged
+- Clean acknowledgment flow: no window operations, just data logging
+- Removed all snooze, flashing, speech, complex timers
+- Basic 30-second refresh timer for tree updates
+
+### Next Steps (Planned)
+- Step 2: Basic acknowledgment without tree rebuild
+- Step 3: Add visual status indicators (colors only)
+- Step 4: Add audio alerts (simple)
+- Step 5: Add window management (careful approach)
+- Step 6: Add system tray (if needed)
+
+## Key Principles for V2
+- **No tree operations during acknowledgments in fullscreen/maximized**
+- **Minimal UI updates - data changes only**
+- **Test each feature before adding the next**
+- **Keep acknowledgment flow completely separate from display updates**
+
+---
+
+## Session Log
+
+### 2025-07-02 - Initial V2 Planning
+- User frustrated with hours of failed attempts
+- Decision to completely rebuild from clean foundation
+- Starting with stripped-back tree UI
+
+### 2025-07-02 - Clean V2 Created and Qt6 Upgraded
+- ✅ Created completely new alert_display.py (220 lines vs 1400+)
+- ⚠️ File became empty somehow - recreated clean version
+- ✅ **UPGRADED TO PyQt6** for better performance and modern features
+- ✅ Updated requirements.txt to use PyQt6
+- ✅ Installed PyQt6 in virtual environment
+- ✅ Application running successfully with Qt6
+
+**V2 with Qt6 Features:**
+- Modern Qt6 framework (better performance, newer APIs)
+- Clean 220-line codebase vs 1400+ broken lines
+- Simple tree structure with time groups and carriers
+- Color coding: Blue=Pending, Red=Active, Dark Red=Missed, Green=Acknowledged
+- Clean acknowledgment flow: no window operations, just data logging
+- Removed all snooze, flashing, speech, complex timers
+- Basic 30-second refresh timer for tree updates
+- Fixed acknowledgment key lookup bug
